\name{biglme-package}
\docType{package}
\title{
Procedure for doing inference of one-way Linear Mixed Effects models.}
\description{
This package aims to make inference of one-way Linear Mixed
    Effects models where the user may specify any non-singular covariance
    structure to the random effects. The procedure also supports big data.}
\details{
\tabular{ll}{
Package: \tab biglme\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-05-26\cr
License: \tab  GPL(>=2) \cr
}
}
\author{
Oyvind Bleka <Oyvind.bleka@fhi.no>
Maintainer: Oyvind Bleka <Oyvind.bleka@fhi.no>
}
\references{
Master Thesis of Oyvind Bleka
}

